Create Navbar, add navbar to tasks page, Improve UI for student task view
