+++ package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
+    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.14.11",
     "@mui/lab": "^5.0.0-alpha.133",
     "@mui/material": "^5.14.18",
