+++ src/pages/Navbar.tsx
@@ -1,162 +1,35 @@
-// Import React and specific hooks (useState and useEffect) from the 'react' library.
-import React, { useState, useEffect, useContext } from 'react';
-// Import components (Typography and Container) from the Material-UI library.
-import { Typography, Container, Button, Paper, Avatar ,Box } from '@mui/material';
-import { Link, useNavigate } from 'react-router-dom'; // Import Link for navigation
-import TAJobDisplayComponent from './TAJobDisplayComponent';
-import { UserContext } from '../provider';
+
+import React from 'react';
+import { Link } from 'react-router-dom';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
 import AvatarWrapper from '../components/AvatarWrapper';
-import AdminDashboard from './AdminDashboard';
-import useAutoLogout from '../components/AutoLogOut';
-import { link } from 'fs';
-import Footer from '../components/Footer';
-import CustomModal from '../components/CustomModal';
 
-// Define an interface 'User' to specify the structure of a user object.
-// interface User {
-//   username: string;
-//   avatarUrl: string;
-//   role: string;
-//   firstName: string;
-//   smuNo: number;
-// }
 
-// Define a functional component called 'Home' using the React.FC (Functional Component) type.
-const Home: React.FC = () => {
-  // Initialize a 'user' state variable using the 'useState' hook, initially set to 'null'.
-  const userContext = useContext(UserContext);
-  if (!userContext) {
-    return <div>Loading...</div>; // or any other fallback UI
+interface NavbarProps {
+    user: User; // Define the type for user
+    handleLogout: () => void; // Define the type for handleLogout
+    handleProfile: () => void; // Define the type for handleProfile
+    handlePerformance: () => void; // Define the type for handlePerformance
+    isHomePage: boolean;
   }
 
-  const { user, setUser } = userContext;
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  // React hooks
-  const navigate = useNavigate();
-
-  //console.log('home:', user);
-
-  if (!user) {
-    return <div>Loading...</div>; // or any other fallback UI
+  interface User {
+    id: number;
+    firstName: string;
+    lastName: string;
+    email: string;
+    username: string;
+    smuNo: string;
+    year: string;
+    avatarUrl: string;
+    role: string;
   }
+  
 
-  const { role } = user;
-
-  const navigateToStudentProfile = () => {
-    navigate('/student-profile');
-  };
-
-  const navigateToFacultyProfile = () => {
-    navigate('/faculty-profile');
-  };
-
-  /**
-   * Log out the user, delete user from localStorage
-   */
-  const handleLogout = function () {
-    localStorage.removeItem('user');
-    setUser(null);
-    setIsLoggedIn(false);
-    navigate('/home-default');
-  };
-
-  const handlePerformance = () => {
-    navigate('/performance-result', { state: { user } });
-  };
-
-  /**
-   * Navigate to the corresponding user profile.
-   */
-  const handleProfile = function () {
-    // Guard clause.
-    if (!user) {
-      return;
-    }
-    // Navigate to student/faculty profile.
-    if (user.role === 'student') {
-      navigate('/student-profile');
-    } else if (user.role === 'faculty') {
-      navigate('/faculty-profile');
-    } else if (user.role === 'admin') {
-      navigate('/admin-profile');
-    }
-  };
-
-  const TIMEOUT_DURATION = 10 * 60 * 1000; // 10 minutes
-
-  const logoutFunction = () => navigate('/login'); // Define your logout action
-
-  // Use the auto-logout hook
-  const { Modal, closeModal } = useAutoLogout({
-    timeoutDuration: TIMEOUT_DURATION, // Use the defined timeout duration
-    logoutFunction: () => {
-      console.log('called');
-      localStorage.clear(); 
-      setUser(null); 
-      setIsLoggedIn(false); 
-      console.log('go to login');
-      navigate('/login'); 
-      
-    },
-  });
-
-  // Use the 'useEffect' hook to execute code after the component renders.
-  useEffect(() => {
-    // Retrieve the 'user' data from local storage, parsing it from JSON, or default to 'null'.
-    const currentUser = JSON.parse(localStorage.getItem('user') || 'null');
-    // Set the 'user' state with the retrieved user data.
-    if (currentUser) {
-      setUser(currentUser);
-      setIsLoggedIn(true);
-    }
-  }, []);
-
-  const renderContent = () => {
-    // When the user is an administrator, display the AdminDashboard component
-    if (user && user.role === 'admin') {
-      return <AdminDashboard />;
-    } else {
-      // Content displayed by non administrator users
-      return (
-        <Container style={{
-          paddingLeft: '10px',
-          paddingRight: '10px',
-          minHeight: '80vh'
-        }}>
-          {/* Large image at the top */}
-          <img
-            src="https://www.smu.edu/-/media/Site/DevelopmentExternalAffairs/MarketingCommunications/digital-marketing/students-hanging-dallas-hall.jpg?h=1333&iar=0&w=2000&hash=EAA3D7A0E96DA001440160E0ECB8643D"
-            alt="SMU Dallas Hall"
-            style={{ width: '100%', height: 'auto', paddingTop: '50px' }}
-          />
-          {/* Text box that spans the page */}
-          {user && (
-            <Paper style={{ padding: '20px' }}>
-              <Typography variant="body1">
-                Welcome to CS5/7328 TA Job Site! This site is for SMU Lyle
-                School of Engineering students to find TA jobs.
-              </Typography>
-            </Paper>
-          )}
-
-          {/* If the user is a student, display their work list */}
-          {user && user.role === 'student' && (
-            <Container maxWidth="sm" style={{ marginTop: '20px' }}>
-              <TAJobDisplayComponent />
-            </Container>
-          )}
-        </Container> 
-      );
-    }
-  };
-
+const Navbar: React.FC<NavbarProps> = ({ user, handleLogout, handleProfile, handlePerformance, isHomePage }) => {
   return (
-    // Render the component within a container with a maximum width of 'sm'.
-
-    <div>
-      {/* Blue banner with "Login" button */}
-      <div
+    <div
         style={{
           backgroundColor: '#1976D2',
           padding: '16px',
@@ -315,10 +188,16 @@ const Home: React.FC = () => {
               ) : (
                 ''
               )}
+              {!isHomePage && (
+                <Button component={Link} to="/home" variant="contained" color="secondary" style={{margin:'10px'}}>
+                  Home
+                </Button>
+              )}
               <AvatarWrapper
                 user={user}
                 onLogout={handleLogout}
                 onProfile={handleProfile}
+                
               />
             </div>
           ) : (
@@ -331,50 +210,13 @@ const Home: React.FC = () => {
             >
               Login
             </Button>
-          )}
+          )
+          }
         </div>
       </div>
-      <div>
-        {/* Call renderContent to display corresponding content based on user roles */}
-        {renderContent()}
-
-        {/* Text box that spans the page, will fill it with about us and stuff BWG */}
-        <Box
-        sx={{
-          py: 2,
-          px: 3,
-          mt: 'auto',
-          color: 'white',
-          textAlign: 'center',
-          position: 'fixed',
-          left: 0,
-          bottom: 0,
-          width: '100%',
-        }}
-      >
-        <Paper style={{ padding: '20px' }}>
-          <Typography variant="body1">
-            Welcome to CS5/7328 TA Job Site! This site is for SMU Lyle School of
-            Engineering students to find TA jobs.
-          </Typography>
-        </Paper>
-        </Box>
-        {/* TODO: hide this Component when user not login */}
-
-        {/* show the TAjob listing if the user is student */}
-        {/* {role === 'student' && (
-          <Container maxWidth='sm' style={{ marginTop: '20px' }}>
-            <TAJobDisplayComponent></TAJobDisplayComponent>
-          </Container>
-        )} */}
-      </div>
-
-      <Footer />
-      {Modal}
-      
-    </div>
   );
 };
 
-// Export the 'Home' component so it can be used in other parts of the application.
-export default Home;
+
+
+export default Navbar;
