+++ src/pages/task/TaskDisplayComponentFaculty.tsx
@@ -2,7 +2,9 @@ import React, { useContext, useEffect, useState } from 'react';
 import api from '../../services/taskform';
 import { UserContext } from '../../provider';
 import { Button } from '@mui/material';
-import { Link } from 'react-router-dom'; // Import Link from react-router-dom
+import useAutoLogout from '../../components/AutoLogOut';
+import { Link , useNavigate } from 'react-router-dom'; // Import Link from react-router-dom
+import  Navbar from '../Navbar';
 interface Task {
   facultyId: number;
   studentId: number;
@@ -14,6 +16,94 @@ interface Task {
 }
 
 const ViewAssignedTasks: React.FC = () => {
+  const userContexts = useContext(UserContext);
+  if (!userContexts) {
+    return <div>Loading...</div>; // or any other fallback UI
+  }
+
+  const { user, setUser } = userContexts;
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  // React hooks
+  const navigate = useNavigate();
+
+  //console.log('home:', user);
+
+  if (!user) {
+    return <div>Loading...</div>; // or any other fallback UI
+  }
+
+  const { role } = user;
+
+  const navigateToStudentProfile = () => {
+    navigate('/student-profile');
+  };
+
+  const navigateToFacultyProfile = () => {
+    navigate('/faculty-profile');
+  };
+
+  /**
+   * Log out the user, delete user from localStorage
+   */
+  const handleLogout = function () {
+    localStorage.removeItem('user');
+    setUser(null);
+    setIsLoggedIn(false);
+    navigate('/home-default');
+  };
+
+  const handlePerformance = () => {
+    navigate('/performance-result', { state: { user } });
+  };
+
+  /**
+   * Navigate to the corresponding user profile.
+   */
+  const handleProfile = function () {
+    // Guard clause.
+    if (!user) {
+      return;
+    }
+    // Navigate to student/faculty profile.
+    if (user.role === 'student') {
+      navigate('/student-profile');
+    } else if (user.role === 'faculty') {
+      navigate('/faculty-profile');
+    } else if (user.role === 'admin') {
+      navigate('/admin-profile');
+    }
+  };
+
+  const TIMEOUT_DURATION = 10 * 60 * 1000; // 10 minutes
+
+  const logoutFunction = () => navigate('/login'); // Define your logout action
+
+  // Use the auto-logout hook
+  const { Modal, closeModal } = useAutoLogout({
+    timeoutDuration: TIMEOUT_DURATION, // Use the defined timeout duration
+    logoutFunction: () => {
+      console.log('called');
+      localStorage.clear(); 
+      setUser(null); 
+      setIsLoggedIn(false); 
+      console.log('go to login');
+      navigate('/login'); 
+      
+    },
+  });
+
+  // Use the 'useEffect' hook to execute code after the component renders.
+  useEffect(() => {
+    // Retrieve the 'user' data from local storage, parsing it from JSON, or default to 'null'.
+    const currentUser = JSON.parse(localStorage.getItem('user') || 'null');
+    // Set the 'user' state with the retrieved user data.
+    if (currentUser) {
+      setUser(currentUser);
+      setIsLoggedIn(true);
+    }
+  }, []);
+
   const userContext = useContext(UserContext);
   const [tasks, setTasks] = useState<Task[]>([]);
   const storedUser = localStorage.getItem('user');
@@ -38,7 +128,17 @@ const ViewAssignedTasks: React.FC = () => {
   }, [storedUser]);
 
   if (!tasks.length) {
-    return <div>No Tasks Available</div>;
+    return (
+      <div>
+      <Navbar
+        user={user} // Pass the user prop
+        handleLogout={handleLogout} // Pass the handleLogout prop
+        handleProfile={handleProfile} // Pass the handleProfile prop
+        handlePerformance={handlePerformance} // Pass the handlePerformance prop
+        isHomePage={false}
+      /> <div>No Tasks Available</div>
+      </div>
+  ) ;
   }
 
   if (!Array.isArray(tasks)) {
@@ -47,6 +147,13 @@ const ViewAssignedTasks: React.FC = () => {
 
   return (
     <div>
+       <Navbar
+        user={user} // Pass the user prop
+        handleLogout={handleLogout} // Pass the handleLogout prop
+        handleProfile={handleProfile} // Pass the handleProfile prop
+        handlePerformance={handlePerformance} // Pass the handlePerformance prop
+        isHomePage={false}
+      />
       <h2>Tasks Assigned</h2> 
       <table style={{ borderCollapse: 'collapse', width: '100%' }}>
         <thead>
