+++ src/stylesheets/App.css
@@ -36,3 +36,5 @@
     transform: rotate(360deg);
   }
 }
+
+
